Modif of CreateCoin - try to do the wind particles
-Now in CreateCoin, the HighestBgPos works for any background with any
number of parts
- Wind particles : two problems
1) the color of the wind : the white/light-grey wind is not very visible
on the background which has bright colors
2) the collision : in Unity 5.2, particles don't collide with
BoxCollider2D, and I can not add a BoxCollider3D to a Rigidbody2D. So I
have added a BoxCollider3D to a sprite behind the Character. Now, the
problem is that a lot of particles don't bounce on the Character and
pass through him.