Created wind effect and attached it to the swipe
DONE :
Several wind prefabs depending on the « force » of the wind : big,
medium, small.
Starting point of the wind is the starting point of the swipe, and same
direction.

TODO :
Can be quite slow : consider speeding up the spawning of particles.
A possible improvement : direct the wind towards the character instead
of just under the swipe of the player.